[libc] remove rtlibc
