+++ bsp/qemu-riscv-virt64/driver/sbi.c
@@ -72,8 +72,8 @@ sbi_get_impl_version(void)
 void
 sbi_print_version(void)
 {
-    u_int major;
-    u_int minor;
+    int major;
+    int minor;
 
     /* For legacy SBI implementations. */
     if (sbi_spec_version == 0)
